refactor: move DisconnectCleanupService to dedicated service package for better code hygiene
