+++ src/test/kotlin/at/aau/hexabrawl/websocketserver/model/DisconnectCleanupServiceTest.kt
@@ -7,6 +7,7 @@ import org.mockito.ArgumentMatchers
 import org.mockito.Mockito.*
 import org.springframework.messaging.simp.SimpMessagingTemplate
 import at.aau.hexabrawl.websocketserver.TestServiceFactory
+import at.aau.hexabrawl.websocketserver.service.DisconnectCleanupService
 
 /**
  * Tests fuer den DisconnectCleanupService.
